ADR-228: Enforce float32 dtype on SpectrogramStage output
librosa.power_to_db may return float64; adding .astype(np.float32) ensures the
saved .npy always matches the documented contract. Added test_output_npy_dtype_is_float32
to verify.

+++ ml/pipeline/speech/spectrogram_stage.py
@@ -106,4 +106,4 @@ class SpectrogramStage(ModifierStage[AudioSample, SampleSpectrogram]):
         else:
             log_s = log_s[:, : self._time_steps]
 
-        return log_s
+        return log_s.astype(np.float32)
